upadate project
Finish table resize.
