+++ src/scss/table.scss
@@ -72,9 +72,10 @@
     opacity: 0;
     position: absolute;
     right: 0;
+    z-index: 1000;
 
     &:hover {
-      opacity: 1;
+      opacity: 1 !important;
     }
   }
